Move parse to gohacks
